requirements.txt
asgiref==3.2.3
Django==3.0.7
mysqlclient==1.4.6
Pillow==7.2.0
pytz==2019.3
sqlparse==0.3.0
channels==2.4.0
channels-redis==3.1.0
selenium
